Wrap setting description to avoid too wide window
The last setting has a very long description and apparently the labels are not wrapping by default. This makes the extension settings window very wide, on my screen it did not fit and made the settings hard to use (need to go into floating mode, then move the window around...).

This sets wrapping so the window can be smaller.